Maintenance
main.lua seems to run before iOS has fully initialized the app.  This causes the UIApplication openURL: call to return false.  Returning false causes Facebook to not use the Facebook app to login, instead it causes the facebook app to use a webdialog.  The login call is now dispatched asynchronously.